loan_management_system/requirement.txt
asgiref==3.8.1
certifi==2024.7.4
click==8.1.7
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
joblib==1.4.0
nltk==3.3
NRCLex==3.0.0
psycopg2==2.9.9
PyJWT==2.8.0
regex==2024.4.16
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
textblob==0.18.0.post0
tqdm==4.66.2
wheel==0.43.0
whitenoise==6.7.0
